move resources into package
