+++ .github/workflows/build_and_test.yml
@@ -52,7 +52,7 @@ jobs:
     - name: Test mu.py
       run: |
         pip install -e .[resources]
-        ./resources/mu/mu.py -o fake/path/mymudet.pkl --plight pl_step_1000 resources/mu/mmc/ice_allm97.pklz
+        resources/mu/mu.py -o fake/path/mymudet.pkl --plight pl_step_1000 resources/mu/mmc/ice_allm97.pklz
         pytest -k test_pdet
 
   results:
